devices: consolidate display/EDID code into shared module
Move display and EDID generation code from virtio/gpu/ into a new
common devices/display/ module, making it available to both the
built-in virtio-gpu and the upcoming vhost-user-gpu implementation.

Changes:
- Add devices/display/ module with three submodules:
- types.rs: EdidParams, PhysicalSize, DisplayInfo and DisplayInfoEdid
(unconditionally available, for both virtio-gpu and vhost-user-gpu)
- edid.rs: ChromiumOS EDID generator (moved from virtio/gpu/edid.rs)
- display_backend.rs: Display backend moved from virtio/gpu/display.rs,
gated behind gpu feature since it depends on krun_display)
- Remove duplicate EdidParams/PhysicalSize definitions that existed in
both display.rs and the old edid.rs; consolidate into types.rs with
a single default of 300 DPI matching the existing virtio-gpu behavior
- Update all import paths from devices::virtio::gpu::display and
devices::virtio::display to devices::display

No functional changes to the built-in virtio-gpu device.
