+++ src/devices/src/lib.rs
@@ -14,6 +14,7 @@ use std::fmt;
 use std::io;
 
 mod bus;
+pub mod display;
 #[cfg(any(target_arch = "aarch64", target_arch = "riscv64"))]
 pub mod fdt;
 pub mod legacy;
