+++ src/devices/src/virtio/gpu/device.rs
@@ -12,8 +12,8 @@ use super::defs;
 use super::defs::uapi;
 use super::defs::uapi::virtio_gpu_config;
 use super::worker::Worker;
+use crate::display::DisplayInfo;
 use crate::virtio::InterruptTransport;
-use crate::virtio::display::DisplayInfo;
 use krun_display::DisplayBackend;
 #[cfg(target_os = "macos")]
 use utils::worker_message::WorkerMessage;
