+++ src/vmm/src/resources.rs
@@ -33,7 +33,7 @@ use crate::vmm_config::net::{NetBuilder, NetworkInterfaceConfig, NetworkInterfac
 use crate::vmm_config::vsock::*;
 use crate::vstate::VcpuConfig;
 #[cfg(feature = "gpu")]
-use devices::virtio::display::DisplayInfo;
+use devices::display::DisplayInfo;
 #[cfg(feature = "tee")]
 use kbs_types::Tee;
 #[cfg(feature = "gpu")]
